Intercept AADSSO new-user creation and instead attach to account. Also output errors

+++ Profile.php
@@ -1,8 +1,9 @@
 <?php
-
 /**
- * Class containing all profile settings and button for linking WordPress account
- * @since  0.2.1
+ * All profile settings and button for linking WordPress account
+ *
+ * @since   0.2.1
+ * @package AADSSO
  */
 class AADSSO_Profile {
 
@@ -38,8 +39,11 @@ class AADSSO_Profile {
 	 * @since  0.2.1
 	 * @return AADSSO_Profile A single instance of this class.
 	 */
-	public static function get_instance( $aadsso ) {
+	public static function get_instance( $aadsso = null ) {
 		if ( null === self::$single_instance ) {
+			if ( $aadsso ) {
+				throw new Exception( 'AADSSO_Profile::get_instance requires the AADSSO instance to be provided' );
+			}
 			self::$single_instance = new self( $aadsso );
 		}
 
@@ -74,9 +78,11 @@ class AADSSO_Profile {
 
 		// Jump in before and after aad-sso to merge the users (if requested)
 		add_filter( 'authenticate', array( $this, 'maybe_cache_user' ), 0 );
-		add_filter( 'authenticate', array( $this, 'maybe_connect_user' ), 2 );
+		add_filter( 'aad_sso_new_user_override', array( $this, 'maybe_connect_before_create' ), 10, 3 );
+		add_filter( 'aad_sso_found_user', array( $this, 'maybe_connect_found_user' ), 10, 2 );
+		add_filter( 'authenticate', array( $this, 'catch_errors' ), 2 );
 
-		add_action( 'all_admin_notices', array( $this, 'maybe_success_notice' ) );
+		add_action( 'all_admin_notices', array( $this, 'maybe_user_notice' ) );
 	}
 
 	/**
@@ -89,7 +95,7 @@ class AADSSO_Profile {
 	public function aad_sso_merge_button() {
 		$user_id = get_current_user_id();
 
-		// if aadsso altsecid, no need to link the account
+		// If aadsso altsecid, no need to link the account
 		if ( get_user_meta( $user_id, $this->aadsso->user_id_meta_key, 1 ) ) {
 			return;
 		}
@@ -126,13 +132,15 @@ class AADSSO_Profile {
 		if ( 'profile.php' != $pagenow ) {
 			$this->maybe_redirect_to_profile( $user_id );
 		} else {
+			$this->maybe_redirect_to_profile( $user_id );
 			$this->maybe_redirect_to_aad( $user_id );
 		}
 
 	}
 
 	/**
-	 * Handles determing if a redirect to the profile screen is needed
+	 * Handles determing if a redirect to the profile screen is needed.
+	 * Will redirect if the profile was linked successfully or if there was an error.
 	 *
 	 * @since  0.2.1
 	 *
@@ -141,9 +149,11 @@ class AADSSO_Profile {
 	 * @return null
 	 */
 	public function maybe_redirect_to_profile( $user_id ) {
-		if ( get_user_meta( $user_id, 'aadsso_is_linked', 1 ) ) {
-			wp_redirect( admin_url( 'profile.php?aadsso_is_linked' ) );
-			exit;
+		foreach ( array( 'aadsso_link_failed', 'aadsso_is_linked' ) as $check_redirect ) {
+			if ( ! isset( $_GET[ $check_redirect ] ) && get_user_meta( $user_id, $check_redirect, 1 ) ) {
+				wp_redirect( admin_url( 'profile.php?'. $check_redirect ) );
+				exit;
+			}
 		}
 	}
 
@@ -172,6 +182,7 @@ class AADSSO_Profile {
 
 	/**
 	 * Handles storing user id to an object property if the conditions are met for merging
+	 * Hooked to 'authenticate', 0
 	 *
 	 * @since  0.2.1
 	 *
@@ -195,25 +206,49 @@ class AADSSO_Profile {
 	}
 
 	/**
-	 * Handles merging an AAD user with the stored user
+	 * Intercept the AADSSO user-creation and instead attach to the user-to-link.
+	 * Hooked to 'aad_sso_new_user_override'
+	 *
+	 * @since  0.2.2
+	 *
+	 * @param  mixed  $null     Return non-null value to override user-creation
+	 * @param  array  $userdata Array of new-user userdata
+	 * @param  object $jwt      JWT object
+	 *
+	 * @return null|WP_User     If we're linking, returns the linked user object
+	 */
+	public function maybe_connect_before_create( $null, $userdata, $jwt ) {
+
+		if ( ! $this->user_to_keep ) {
+			return $null;
+		}
+
+		$user = new WP_User( $this->user_to_keep );
+
+		// Stop recursion
+		remove_filter( 'aad_sso_found_user', array( $this, 'maybe_connect_found_user' ), 10, 2 );
+
+		return $this->connect_accounts( $user, sanitize_text_field( $jwt->altsecid ), true );
+	}
+
+	/**
+	 * Handles merging an AAD user with the stored user.
+	 * Hooked to 'aad_sso_found_user'.
 	 *
 	 * @since  0.2.1
 	 *
-	 * @param  mixed  $user User object or WP_Error
-	 *
-	 * @return mixed        User object or WP_Error
+	 * @return mixed User object or WP_Error
 	 */
-	public function maybe_connect_user( $user ) {
+	public function maybe_connect_found_user( $user, $jwt ) {
+
 		if (
-			$this->user_to_keep
-			&& is_a( $user, 'WP_User' )
-			&& ( $aad_sso_id = get_user_meta( $user->ID, $this->aadsso->user_id_meta_key, 1 ) )
+			! $this->user_to_keep
+			|| $this->user_to_keep == $user->ID
 		) {
-			$user->aad_sso_id = $aad_sso_id;
-			return $this->connect_accounts( $user );
+			return $user;
 		}
 
-		return $user;
+		return $this->connect_accounts( $user, sanitize_text_field( $jwt->altsecid ) );
 	}
 
 	/**
@@ -221,26 +256,36 @@ class AADSSO_Profile {
 	 *
 	 * @since  0.2.1
 	 *
-	 * @param  WP_User  $user_to_link New WP_User object to link existing WP user
+	 * @param  WP_User $user_to_link New WP_User object to link existing WP user
+	 * @param  string  $aad_sso_id   AADSSO user id
+	 * @param  bool    $same_user    If getting a user from maybe_connect_before_create, the user will be the same.
 	 *
-	 * @return WP_User                The stored user object after merge
+	 * @return WP_User               The linked user object after merge
 	 */
-	public function connect_accounts( $user_to_link ) {
+	protected function connect_accounts( $user_to_link, $aad_sso_id, $same_user = false ) {
+
+		$updated = update_user_meta( $this->user_to_keep, $this->aadsso->user_id_meta_key, $aad_sso_id );
+
+		if ( $same_user ) {
+			do_action( 'aad_sso_link_user', $user_to_link->ID, $this->user_to_keep, $same_user );
+		} else {
+			do_action( 'aad_sso_link_users', $user_to_link->ID, $this->user_to_keep, $same_user );
+		}
+
+		if ( ! $same_user ) {
+
+			// WordPress User Administration API
+			require_once( ABSPATH . 'wp-admin/includes/user.php' );
+
+			$this->reassign_comments( $user_to_link->ID );
+			wp_delete_user( $user_to_link->ID, $this->user_to_keep );
+
+		}
 
 		delete_user_meta( $this->user_to_keep, 'is_aadsso_linking' );
 		update_user_meta( $this->user_to_keep, 'aadsso_is_linked', 'true' );
-		update_user_meta( $this->user_to_keep, $this->aadsso->user_id_meta_key, $user_to_link->aad_sso_id );
 
-		// WordPress User Administration API
-		require_once( ABSPATH . 'wp-admin/includes/user.php' );
-
-		$this->reassign_comments( $user_to_link->ID );
-
-		do_action( 'aad_sso_link_user', $user_to_link->ID, $this->user_to_keep );
-
-		wp_delete_user( $user_to_link->ID, $this->user_to_keep );
-
-		return get_user_by( 'id', $this->user_to_keep );
+		return $same_user ? $user_to_link : get_user_by( 'id', $this->user_to_keep );
 	}
 
 	/**
@@ -271,18 +316,73 @@ class AADSSO_Profile {
 	}
 
 	/**
-	 * Determines if a merge-success message should be displayed
+	 * If there was an error during linking, save to user-meta for later output
+	 * Hooked to 'authenticate', 2
+	 *
+	 * @since  0.2.2
+	 *
+	 * @param  WP_User|WP_Error $user User object or WP_Error object
+	 *
+	 * @return WP_User|WP_Error       User object or WP_Error object
+	 */
+	public function catch_errors( $user ) {
+		if ( $this->user_to_keep && is_wp_error( $user ) ) {
+			update_user_meta( $this->user_to_keep, 'aadsso_link_failed', $user->get_error_messages() );
+		}
+
+		return $user;
+	}
+
+	/**
+	 * Determines if a merge-success or error message should be displayed
+	 * Hooked to 'all_admin_notices', 2
 	 *
 	 * @since  0.2.1
 	 *
 	 * @return null
 	 */
-	public function maybe_success_notice() {
+	public function maybe_user_notice() {
+		if ( isset( $_GET['aadsso_link_failed'] ) ) {
+			$this->maybe_error_notice();
+		}
+		if ( isset( $_GET['aadsso_is_linked'] ) ) {
+			$this->maybe_success_notice();
+		}
+	}
 
-		if ( ! isset( $_GET['aadsso_is_linked'] ) ) {
+	/**
+	 * Determines if a merge-fail message should be displayed
+	 *
+	 * @since  0.2.2
+	 *
+	 * @return null
+	 */
+	public function maybe_error_notice() {
+		$user_id = get_current_user_id();
+
+		$errors = get_user_meta( $user_id, 'aadsso_link_failed', 1 );
+
+		if ( ! $errors ) {
 			return;
 		}
 
+		if ( is_array( $errors ) ) {
+			$errors = count( $errors ) > 1 ? '<ul><li>'. implode( '</li><li>', $errors ) . '</li></ul>' : end( $errors );
+		}
+
+		echo '<div id="message" class="error"><p>' . sprintf( __( '<strong>ERROR:</strong> %s', 'aad-sso' ), $errors ) . '</p></div>';
+
+		delete_user_meta( $user_id, 'aadsso_link_failed' );
+	}
+
+	/**
+	 * Determines if a merge-success message should be displayed
+	 *
+	 * @since  0.2.2
+	 *
+	 * @return null
+	 */
+	public function maybe_success_notice() {
 		$user_id = get_current_user_id();
 
 		if ( ! get_user_meta( $user_id, 'aadsso_is_linked', 1 ) ) {
@@ -292,7 +392,6 @@ class AADSSO_Profile {
 		echo '<div id="message" class="updated"><p>' . __( 'Success, your profile has been linked to your Azure Active Directory account.', 'aad-sso' ) . '</p></div>';
 
 		delete_user_meta( $user_id, 'aadsso_is_linked' );
-
 	}
 
 	/**
